+++ apps/api-server/src/routes/api/lock.js
@@ -3,7 +3,7 @@ const createError = require('http-errors');
 const db = require('../../db');
 const auth = require('../../middleware/sequelize-authorization-middleware');
 const pagination = require('../../middleware/pagination');
-const searchResults = require('../../middleware/search-results-static');
+const searchInResults = require('../../middleware/search-in-results');
 
 const router = express.Router({ mergeParams: true });
 
@@ -34,7 +34,7 @@ router.route('/$')
       .catch(next);
   })
   .get(auth.useReqUser)
-	.get(searchResults)
+  .get(searchInResults({ searchfields: ['type'] }))
 	.get(pagination.paginateResults)
 	.get(function(req, res, next) {
 		res.json(req.results);